I modified the output to make it easier to parse.
I modified the output to make it easier to parse.  Basically, with the
way things are right now, the output is extremely easy to parse with
utilities like awk.  In addition, I started moving around the examples
and modifying the source in order to make it easier to write the manual.
Finally, here's the first bit of the manual, but it's far from complete.