I attached the copyright and license notice to...
I attached the copyright and license notice to the core files.
Basically, the entire project falls under the copyright and license in
the main directory.  However, these core files are the files most likely
to be removed and included into a code by hand.  In that case, the
copyright notice, while still legally in affect, isn't visible.  Just
adding a comment with the notice fixes that.