I reworked the line-search pieces of the code...
I reworked the line-search pieces of the code so that non-Newton methods
works far more efficiently.  The difficulty here is in how to adjust the
line-search parameter alpha.  Basically, I added precondioned versions
of the steepest descent and nonlinear-CG codes.  Then, I ran them on a
quadratic problem with a near perfect preconditioner.  In this setup,
the problems should converge in exactly the same number of iterations
that CG does since nonlinear-CG=CG for quadratic problems.  However,
this was not the case.  What was happening was the code was unable to
compute a near exact line-search, which was necessary for the
equivalence to hold.  Previously, the line-search parameter was
initialized to something the user gave.  Then, during the line-search,
we searched at 2 times this parameter and headed back toward zero.  The
exact process depended on whether or not we were using a golden section
search or a backtracking linesearch.  Then, the alpha found was used as
the guess for the next search.  The problem with this is that
nonlinear-CG was taking a short step on the first iteration and then
unable to take a longer step on the second iteration since 2*alpha was
not of sufficient length.  This led to a longer puzzling over whether or
not using the last found alpha was necessarily the best alpha to use for
the next linesearch.  In order to resolve this issue, I made the
following algorithmic change.  Now, we have a base line-search
parameter, alpha0.  This is the parameter that is now set by the user in
the json file.  The line-search will *not* search any farther than
alpha0.  Basically, the line-search is being used as a safe-guard for
convergence (and making sure we know exactly how far out we search is
very important for the interior point method lest we violate the
inequality).  Then, we find the actual line-search parameter alpha,
which is guaranteed to be less than or equal to alpha0.  By starting our
line-search at alpha0 at every iteration, we can be smarter about how we
dynamically adjust alpha0.  For the golden section search, if we
continually bracket to the right, we probably need to increase alpha0.
Alternatively, if we continually bracket to the left, we probably need
to decrease alpha0.  However, if we find a parameter somewhere in the
middle that we're happy with, then our current alpha0 is probably ok.
Next, I modified the backtracking line search to essentially do no work
and simply try to take a full step at alpha0.  If this step is good, we
take it, if not, we use the safeguarding procedure built into the line
search to backtrack.  In addition, we now satisfy the sufficient
decrease condition in the line search method.  This is tuned by a new
parameter c1.  Hence, we now guarantee that every step satisfies
f(x+alpha dx) < f(x) + c1 alpha <grad f(x),dx).  In the future, it
should be pretty straightforward to add the curvature condition, but I'm
not entirely convinced we want that now (it requires extra gradient
computations).  Outside of modifying the line-search procedure, I added
preconditioned versions of steepest descent and the nonlinear-CG
variations Fletcher-Reeves, Polak-Ribiere, and Hestenes-Stiefel.  These
will use the preconditioner defined by PH_type.  If this is UserDefined,
then it will look to PH for the preconditioner.  It may be good to put
this somewhere else, but we really do want an approximation to the
inverse of the Hessian, so it does make some sense to put that code
there.  Finally, I added and updated tests for all of these
combinations.  As a note, as opposed to our last line-search code, the
current setup appears to be much more efficient than before.