I improved the parameter verification to make...
I improved the parameter verification to make sure that we now have
enough line-search iterations.  This also means that we should no longer
be allowed to use the golden-section search with fewer than two
iterations.  The origin of this was that I was previously reporting five
line-search iterations, but we were doing seven objective evaluations.
Basically, I neglected to account for the two objective evaluations that
established the bracket.  This has been fixed meaning that if the LSIter
column says five, it really did five objective evaluations for the
line-search.  In addition, I set some flags to insure that the
Barzilai-Borwein two-point line-search can only be used with the
steepest descent direction.  Really, their algorithm is the direction
and the line-search, but we do things in this way to shoe horn the
algorithm into the code.  In any case, the change to the line-search
required a rebaselining of many of the unit tests.  Finally, I fixed the
formating of the BSD license.