I use a trick when integrating the interior...
I use a trick when integrating the interior point code with the equality
constrained code where I communicate how much a step needs to be
truncated via the line search parameter.  Really, we need to do it this
way so that we know when to update the Lagrange multiplier.  Anyway,
after I modified the code with the new line-search scheme with alpha0
and alpha, I forgot to update this piece.  This check-in fixes that.