Note: This check-in and those going forward...
Note: This check-in and those going forward are under copyright from
OptimoJoe.  Everything is still BSD licensed, but make sure to use the
new copyright notices.  Outside of that, I updated the code to work with
C++11 constructs.  Basically, the problem we had before with the memory
management is that we required a default constructor for the vectors,
but then we called the copy constructor on the object initialized from
the default constructor in order to get it into a stl::list.  We really,
really don't want to call the user defined copy constructor since it
could do all sorts of odd things in a scientific code.  By using
emplace_back rather than push_back, we can avoid calling the
copy_cosntructor.  Next, I moved all of our auto_ptrs to unique_ptrs.
These are the new standard for memory management, so it's good to finish
that port.  In addition, I removed the references to drand48 since this
only worked on POSIX systems.  On Windows, the functions didn't exist
and I needed to use a macro.  Now, we use the new STL random number
bits.  For reference, this is needed to choose an initial Krylov vector
for our eigenvalue solvers.  Finally, I need to do some cleanup on the
Matlab interfaces.  Right no, I backed off some of the exception
handling in order to make sure things worked with Octave.  Octave
doesn't have the routines for exception handling that Matlab does, so
I'll probably end up doing more error checking by hand.