I added safe defaults for ared, pred, and alpha.
I added safe defaults for ared, pred, and alpha.  The reason for this is
is that, depending on the algorithm, the restart files contain NaNs,
which stops the restart from occuring properly.  That being said, many
values are initialized to NaN, but these should all be overwritten on
the first iteration.  If this becomes not the case, we should revisit
this.