This is a pretty major overhaul in which I...
This is a pretty major overhaul in which I eliminated the dependency on
boost unit test and converted everything over to ctest, within the cmake
suite.  For some of the tests, like the linear algbra tests, this was a
straight forward port.  For the optimization spefic tests, I now test
directly on the examples.  I accomplish this by using the ability to
generate and diff restart files and then comparing to a directory full
of tests.  This, in theory, should make it much easier to generate new
unit tests based examples and prevent duplicate code between the unit
tests and examples.  Now, that being said, there are still some issues.
As it turns out, json can not represent ieee arithmatic properly.
Specificallly, it can not handle nans and infs.  This causes problems of
the restart files since it's very plausible that a parameter will be a
nan for an iteration.  I massaged the unit tests so this doesn't come up
at the moment.  In the longer term, I need to figure out a resolution
for this problem.  Of course, there were some sort of dumb nans being
generated that didn't hurt anything due to strict versus nonstrict
tolerances.  This occurred in the linear algebra routines, which has
since been fixed.  In addition, the serialization helped find an
outstanding but in peopt.h