I modified the project in order to remove the...
I modified the project in order to remove the header only installation.
At issue was that the trick I used in the past to accomplish this was
not stanards compliant and only worked on Mac and Linux.  It did not
work on Windows.  More specifically, if we look at the unconstrained
code, we see that we called the truncated_minres routine.  This routine
lies on a number of BLAS and LAPACK routines such as the givens
rotations.  Now, we have templated implementations of these, at least
for unconstrained and equality constrained problems, but, in general, we
want to use the optimized BLAS and LAPACK versions.  For this to be
correctly defined, we need to denote the fact that we have optimized
versions in the linalg.h header by providing the prototype for the
template specialization.  However, if we do this, we must link the
library as these versions are in the linalg.o object file.  Now, on
Linux, these BLAS and LAPACK symbols were just weak symbols.  Hence, if
we linked the optimized version, they were overridden.  However, if we
did not link optimized versions, we got the templated versions.  On
Windows, these were not weak symbols, but actualy symbols, which caused
an issue.  Technnically, Windows was correct in this case.  We should
have declared the prototype that specialized these functions.  But,
again, this requires linking the library.  Once we go down this road, it
makes sense to remove bits and pieces out of peopt.h that could be put
into an implementation file.  Before, we kept playing a game with inline
functions to get the header only mode.  Finally, due to this new build
strategy, when we link, we will likely always need to link BLAS and
LAPACK in our projects.  We do not need to link the jsoncpp library
unless we use jsoncpp features.