I reworked the build process for the Python...
I reworked the build process for the Python library.  Previously, I
compiled all the python bits in to the shared library for peopt.  I
don't really think it should go there, so I changed it so that Python
builds its own version of the library.  The slightly irritating thing
about this is that the new file peopt_python.pylib needs to be in the
LD_LIBRARY_PATH on Linux and probably something similar on Windows.
Also, it looks like there may be a way to eliminate peopt.py and roll
everything into the shared library.  This is worth a look in the future.
In addition, for whatever reason, it looked like the Matlab build was
broken, so I fixed that.  Finally, notice that I'm using this odd name
.pylib for the shared library.  Basically, on Linux shared libraries are
named .so and on windows they are .dll.  In peopt.py, I need to load in
the shared library by name, which requires, as far as I know, the
extension.  That's some of the reason why it'd be nice to eliminate the
peopt.py file all together and roll it into a shared library.