It turns out the indexing that jsoncpp uses...
It turns out the indexing that jsoncpp uses for it's Value types is not
64-bit safe.  As such, when compiling on 64-bit platforms, we had an
ambiguity because we had operator[] defined for int and ArrayIndex,
which is just unsigned int.  On our side, we index with Natural, which
is set to be size_t.  Generally, this is an unsigned long on 64-bit
platforms.  In any case, the compiler got confused as to which function
to use, so now we force a cast to ArrayIndex.  In theory, in the future,
we may be able to modify jsoncpp so that ArrayIndex=sizet.  Until that
time, don't use the built-in serialization to json for vectors requiring
more than a 32-bit index.  Given the file format, this was probably a
bad idea anyway.  Note, this is *not* a limitation of our restart
procedure, but rather our default way of serializing with jsoncpp.