This is a pretty big checkin where I switch...
This is a pretty big checkin where I switch everything to our new name
Optizelle.  For naming conventions, I try to keep Optizelle in upper
case everywhere except the file names.  In some places, we end up with
Optizelle.py, instead of optizelle.py, in order to make sure that the
Python calls use Optizelle.  Other than that, we try to also standardize
on .param for parameter files and .rst for restart files.