I modified the setup where we consistently use...
I modified the setup where we consistently use the name  .json instead
of .param or .rst when we are using json formatted files.  Frankly, it
makes syntax hilighting much easier for me and the users.  In addition,
I simplified the unit test setup because I combined the parameters as
well as the restart information into a single json formatted file.