This is a big change over of how the Python...
This is a big change over of how the Python interface works.  In the
past, we had a somewhat restricted interface that did not allow access
to things like the StateManipulator or the full optimization state after
optimization completed.  In some sense, this old interface was much
easier to design and implement because a fuller interface requires add
subtleties such as integrating the Python inheritance system with C++.
In this new change, all of this has been worked out.  At this point,
there's probably a few missing functions such as the restart
functionality and the diagnostics.  I'll add these pieces in shortly.
Finally, for the default, provided vector space, which didn't exist in
the past, I'm using numpy.  Hence, we now have a soft dependence on
numpy.  If the user doesn't want this, they're free to implement their
own vector spaces and the numpy example should provide a template for
this.