This is a pretty big change that will...
This is a pretty big change that will temporarily break the Python and
Matlab wrappers.  At its core, it modifies the requirements of the
vectors in the vector space.  Previously, when we were using pre C++11
constructs, we actually required the default constructor as well as the
copy constructor to be available.  As it turned out, the copy
constructor was *only* called on the default constructor, but it was
there.  In addition, we used std::lists to store all of our variables.
The reason for this was that by using the splice function from
std::list, we could essentially create a poor man's move semantics.
With C++11, we can actually use proper move semantics.  This allows us
to remove the std::list trick and actually store real variables in the
state.  This also allows us to change the requirements on the Vector
space.  Now, we no longer require the copy constructor or the default
constructor.  We require only the move constructor and the move
assignment operator.  The latter seems to only be used by the restart
mechanism.  In addition, this modifies the init function inside the
vector space.  Now, it has a signature Vector init(Vector & const x).
Basically, it takes an x, initializes some memory and then moves it
back.  In theory, we could rely on the return value optimization (RVO),
but this makes me a little nervous, so I tend to call std::move in order
to force the memory to be moved.  Finally, I tried to do a once over of
the code to fix some messed up indenting and fix our const correctness.
In general, I try to make as many things constant as possible and I use
the more C++ish syntax Type const * const or Type const & in order to
make it easier to figure out what in the world is constant.  Previously,
I had stuff like const Type &, which was correct.  At least for the
moment, I find it easier to keep things straight by writing them
backwards.