This fixes the compatibility with the new...
This fixes the compatibility with the new version of Optizelle with
Matlab and Python.  Honestly, I'm not happy with the Matlab design right
now, so I need to modify things so that it's similar to Python.
However, in any case, the old bits of code that use the Matlab wrapper
should work correctly.  Essentially, I modified both the Python and
Matlab wrappers to correctly use things like the move semantics required
by Optizelle in addition to the new setup for the vectors.