I'm trying to clean up the json serialization...
I'm trying to clean up the json serialization and deserialization bits.
The deserialize function was slightly off from the rest of the code in
the sense that deserialize really looks like an init function.  In that
sense, it needs to allocate memory and then return a vector, which can
then be moved into the correct place.  In other words, it needs to look
similar to the init function in the vector space, but instead of being
given a vector to initialize from, it's given a json tree.  I've made
these changes and then adjusted our unit tests and examples accordingly.