I modified the serialization procedure again...
I modified the serialization procedure again in an attempt to make the
process of porting this to Python and Matlab easier.  Now, the serialize
routine accepts a vector and returns a string that contains a json
representation of the vector.  Conversely, deserialize accepts a json
string representation of a vector and returns a vector.  Like before,
this works like the init function.  Before, we were modifying jsoncpp
structures directly, which was convenient, but was going to be hard to
port directly to other languages without wrapping all of jsoncpp.  With
this change, we largely isolate ourselves, but it does mildly complicate
the internal code since we're converting back and forth between jsoncpp
objects and json strings when we use the serialize and deserialize
routines.