I've continued to modify the restart mechanism...
I've continued to modify the restart mechanism and have the first pieces
of an implementation for Python.  For the most part, it seems to work
ok.  I kind of want to change the type of the restart packages like
X_Vectors, Reals, Nats, etc. from ([],[]) to [(,)].  Originally, we had
a tuple of a lists because that was the only way to safely move memory
from the vectors to the restart mechanism without copying memory.
Recall, we used to use a poor man's move semantics by playing games with
splice.  Now, since we have actual move semantics, this may not be
necessary.