This is another big check-in.
This is another big check-in.  I modified the restart mechanism so that
we no longer have a tuple of lists, ([],[]), but now have a list of
tuples, [(,)].  The first element in the tuple is the label of the item
and the second element is the item itself.  Long ago, I decided to work
with a tuple of lists because I didn't have move semantics.  Basically,
all of the variables were stored in lists because we could use the
splice routine to move memory around without using copies.  In order to
move the memory into the restart with splice, we needed a pure list to
work with.  Now, at the time, this sort of irritated me, but it worked.
Now, since we have real move semantics, it's possible to switch to the
alternative, a list of tuples, which is what I've done.  In truth, it
was sort of a pain and there may be some bugs.  That being said, I
actually did fix a few outstand bugs related to the restart where
certain elements were not being copied out.  By calling move semantics
on everything, including things like naturals and reals, it caused some
bugs to become more visible, which is good.  In addition, I completed
the restart code inside of the Python wrappers.  In order to get this
done, I did a bit of a refactor, which was also uglier than what I'd
like.  Right now, there's a horde of routines that convert back and
forth between C++ and Python and I was having trouble keeping track of
them.  For example, I overloaded functions like setVector to both
convert to and from Python depending on the argument order.  In theory,
this isn't a bad idea, but it was confusing me, so I just created to big
name spaces called toPython and fromPython to keep things straight.  The
problem with doing this is that it's hard to have good names inside of
the namepsaces, so I just used basic names.  This caused no end of name
conflicts, so there's all sorts of annotations to make sure I get the
right thing named Vector, etc.  Now, the compiler will catch these
problems, but it still is a bit irritating.  However, using these
functions elsewhere in the code has been nice.  There's probably a
better way to organize things, but I'll look at that later.