I completed the reading and writing json...
I completed the reading and writing json formatted restart files for
Python.  This required some changes to how the json restart files are
used.  Previously, I tried to setup deserialize so that it worked like
the init function and correctly allocated memory.  For basic vectors,
this wasn't a problem, but it was an issue when we interfaced in Python.
Basically, the json parser doesn't have any information about the vector
space functions, which are attached to the vectors themselves since we
use a generic Python vector space.  As such, even though we correctly
parsed the data for the vector, we didn't have all the information to
make it complete.  A similar situation arises when using something like
an MPI vector.  With MPI, we need to know the communication structure
and this is generally setup when the program runs.  There's not really a
good way to learn of this information by purely parsing a JSON file
(that I know of.)  In response, I modified the deserialize routine to
accept an additional argument, a constant vector, that is assumed to be
correctly initialized.  We don't modify this vector, but we do use it to
allocate a new vector, which is then returned.  In this way, we're able
to grab any additional, missing information.