I reworked the json parser so that we better...
I reworked the json parser so that we better handle things like NaNs and
Infs.  Basically, NaNs and Infs are not part of the json standard.
Although jsoncpp will output something like nan if it detects an invalid
number, it does not parse these quantities correctly.  This was a
problem in the past, which forced me to set defaults for values in the
state that I didn't really want to.  In any case, I modified the parser
so that it if it detects a NaN or an Inf it will output a string with
this information.  Conversely, when it reads a valid, it will
automatically convert these strings to the correct value.  In theory,
this should make writing some unit tests a bit easier since, in the
past, we really need the first iteration to flesh out these values.
Now, it shouldn't matter.