With this check-in, I improved the function...
With this check-in, I improved the function diagnostics and finite
difference tests.  Previously, if we wanted to do something like a
finite-difference test, we computed them by hand with the functions in
the namespace Diagnostics.  This makes sense in that information about
functions has nothing to do with optimization, so these diagnostic
functions were separated from the optimization routines.  However, in
practice, this is really irritating.  Frankly, this is an optimization
package and we normally run these diagnostics when something goes wrong.
As such, rather than having to modify the code to add the proper
functions, it is preferable just to modify the json file and get the
tests.  With this check-in, this is now possible.  Of course, there are
some minor caveats.  Foremost, in order to enable any test at all, the
variable dscheme (for diagnostic scheme), must be modified.  This will
either fire off the diagnostics at the beginning of the optimization and
then quit or run the diagnostics at every iteration.  Then, the
variables f_diag, g_diag, and h_diag must be modified.  By default, they
compute no diagnostics (NoDiagnostics).  In order to change that, they
can be set to either FirstOrder or SecondOrder to compute either the
first order diagnostics or the first and second order diagnostics,
respectively.  Yes, SecondOrder also does the stuff in FirstOrder.  Part
of the reason for this is that I really don't want to add a third
option.  Also, generally, if you're testing second order information,
you probably need to test the first-order as well.  In any case, the
typical use will be for dscheme=DiagnosticsOnly and
f_diag,g_diag,h_diag=SecondOrder.  This will run the diagnostics and
then quit.  Outside of that, it looks like the Python stuff is pretty
near parity with the C++.  As such, I should start working on unit tests
for all of the Python.