I made a change to the vector space structure.
I made a change to the vector space structure.  Now, all vector spaces
require the function rand, who accepts a vector x and randomizes it.
This is something that I've been contemplating for quite some time.  For
optimization, we don't need this function.  However, it's something that
comes up often when doing diagnostic tests both on the functions as well
as the vector spces.  In the past, I've always had the opinion that
things needed to be minimalist.  If the user needed random variables,
they could make their own and call the diagnostic functions on these
vectors by hand.  In practice, this actually ended up irritating me more
than anything.  Really, I hated having to modify my code everytime
something would break just so I could have a good diagnostic test.  The
first step to fixing this was to rework the finite difference tests so
that they could be run by simply changing the json file.  In this former
change, the direction we would use for the test would be at x, in the
direction dx, dx_old, dy, and dz.  The problem with this scheme is that
I rarely initialize dy and dz to something other than zero.  If they're
zero, we have a very uninteresting finite difference test.  Hence, I was
back to having to insert test code into all of my programs when they
broke.  Really, it's just easier to have a rand function lying around
since it's not that arduous to have in the vector space.  This change
will also likely make it easier to add other diagnostic tests such as on
the vector space functions.  These tests are important since this is a
constant source of error for me as well as my users.  As few minor
notes, I fixed the optimization loop so that we don't have to play an
odd hack to stop the optimization from running when dscheme is set to
DiagnosticsOnly.  Simply, it doesn't run, which is desirable.  Also,
note, all tests still use the point x to run the tests at.  This is
generally what we want when we have an error.  Likely, we find a point
where everything breaks, so we're checking it.