Again, I simplified the build setup with some...
Again, I simplified the build setup with some additional macros.  Now,
we should just need to call a few macros to correctly build and install
the examples or unit tests.  The goal with these new changes is to make
it easy to add Python examples since they should be sharing supporting
files with the C++ examples.