I completed the change over for all of the...
I completed the change over for all of the Python examples and unit
tests.  Now, for the most part, we should have parallels for the C++
examples and unit tests from Python.  What's missing are examples in
Python that need symmetric cones (sdp and socp).  I need to write that
vector space for Python still.  In addition, the
simple_infeasible_inequality example found a bug in Optizelle through
the Python test.  Basically, if we're feasible, the calcualtion for the
Cauchy point in the quasi-normal step divides by zero, which gives us a
bunch of NaNs and messes up the algorithm.  Technically, I've added
enough safeguards so that the optimization will eventually terminate.
It looks like this was happening for C++ case, but it was taking long
enough in the Python case (if you haven't guessed, Python is slower than
C++), that I was able to see that something was wrong.  The solution
that I added was to check if we're feasible prior to the quasi-norm step
calculation.  If we are, which is defined by ||g(x)|| <= eps_constr ||
g(x_typ) ||, we just don't do the quasi-normal step (set the Cauchy
point and step to zero).  In some sense, this does make sense.  If we're
feasible, then why bother with the quasi-normal step?