I modified things so that now all of the...
I modified things so that now all of the examples are actually run as
unit tests.  Prior to this, I wasn't running the diagnostic_checks or
rosenbrock_custom_VS examples regularly, which was a mistake since they
both give an example of and run important features.  As part of this, I
needed to create yet some more macros to handle these tests.  In truth,
my handling of the naming for these CMake macros is terrible.  I'm sure
there's a better way, ut they'll have to suffice for now.  In addition,
I was getting a segault in the finite difference tests that were
triggered by the Python version of the diagnostic_checks example related
to string handling.  It turns out, I had a different problem (the
equality multiplier was the wrong size), but it also seemed like my
handling of the strings could have been better, so I modified things.
Bascally, I was using + for string concatonation, when I probably would
have been better served to use stringstream.  It's minor, but it helped
me find the other error, so I'll keep it.