I modified the specification for Operators,...
I modified the specification for Operators, ScalarValuedFunctions, and
VectorValuedFunctions, StateManipulators, and GMRESManipulators.
Before, we required the user to implement the operator () for the
evaluation.  The problem with this is that the other language wrappers
don't necessarily support these operators.  Instead, we've been using a
function called eval.  This change converts all of the operator () to
eval for consistency.  In some sense, this isn't too big a deal since we
were already called some function called grad, hessvec, p, ps, or pps
for the other function properties.  It looked sort of slick to just
right f(x) instead of f.eval(x), but only for these cases.  As part of
this, I was having trouble getting the StateManipulators and
GMRESManipulators consistent.  The problem is that unlike the function
specifications, the eval function was not pure virtual.  This was
because we need to be able to create empty manipulators for the base
case.  However, by having empty eval functions that are virtual, if we
screw up on our specification, we really don't know until our unit tests
break.  This has happened to me before, so I decided to fix things
better this time.  Now, the eval function is purely virtual, but I
create two extra manipulators called EmptyManipulator and
EmptyGMRESManipulator that do nothing.  This makes everything very
explicit and will hopefully avoid issues in the future.  Finally, I
eliminated the unit test derivative_check since the example
diagnostic_checks really runs over these examples.