This is an intermediate check-in of the Matlab...
This is an intermediate check-in of the Matlab code.  For the most part,
it follows the same pattern as the Python, but it's slightly different.
Namely, we don't have a way to pass things by reference in Matlab, so we
have to modify how functions return values.  Ostencibly, this doesn't
cause any problems, but it would be nice to have some sort of generic
wrapper interface that we build the Python and Matlab interfaces from
and this may cause problems.  Though, it's something to resolve later.
The modifications to the Python code fix minor things like spacing.  I
also forgot to check some of the new vector space functions that were
added.