Another intermediate check-in where, in theory,...
Another intermediate check-in where, in theory, the unconstrained code
for Matlab is complete.  For most of the functions to convert to and
from matlab, there's no actual change from the Python code, which tells
me that there's a fair amount that can be abstracted.  On the Python
front, I made one minor change from PyInt_FromSsize_t to
PyInt_FromSize_t.  Personally, I prefer the latter since our Naturals
are not signed.  What's strange is that going the other way, we only
have PyInt_AsSsize_t and not PyInt_AsSize_t, which is slightly
irritating.