This should give a working version of the new...
This should give a working version of the new Matlab wrapper for
unconstrained problems.  As before, most of the code is similar to what
needs to happen in Python, but there are a couple of variations due to
typing and memory management that are different.  For example, the
serialization routines can't play the registration class trick like
Python.  Also, it's much more difficult to check the type of something
in Matlab than it is in Python because the class and typing systems are
less developed.  On top of this, there are a couple of irritating
catches when returning variables from a mex file.  We can see this in
functions like the restart routines.  Basically, we want to be able to
return cell arrays, but we can't grow such objects on the fly.  We must
know their size a priori, which we don't until we actually do the
Optizelle restart.  Finally, when we work our way through a Python
structure and extract elements, Python allocates new memory for these
objects.  In Matlab, it does not.  As such, we must be *very* careful
not to free memory where we should not.  On the Python side, I noticed
there was a bunch of redundant code in the restart routines where we
check if we have valid restart packages (lists of items and names).
This has been abstracted out.  Finally, I modified the rosenbrock
example, so that it should work properly with the new wrappers in
Matlab.