I added some additional error checking to the...
I added some additional error checking to the serialize routines in
Python.  It's sort of awkward to having error checking in these routines
because we don't have a messaging object.  As such, if there is an error
there, I just throw the system error processing errors.  This is the
same thing that I do for Matlab.