I added the equality constrained bits to the...
I added the equality constrained bits to the Matlab wrapper.  In
addition, I readded the parameter estimation example for Matlab that
shows how to write a code that simultaneously generates both a reduced
space as well as a full space formulation of a parameter estimation
problem.  At the moment, I don't run this example as part of the unit
tests since the problem is randomly generated and not always solvable.
In any case, I think it's a good example of how flexible the framework
is and how little difference there can be between full and reduced space
formulations.