I modified how Optizelle is loaded into Matlab.
I modified how Optizelle is loaded into Matlab.  Before, the functions
setupOptizelle returned a structure that contained all of the funtions
and enumerated types required in order to run a problem.  This isn't
very Matlab-like, but it matches up much closer to how C++ an Python
optimization instances are created, which gives us better consistency
across languages.  Anyway, I actually ended up calling this function
from inside the mex file in order to get access to the enumerated types.
As it turned out, this access was slow.  I speeded things up slightly by
using a persistent variable to not recalculate the enumerated types, but
the access was still slow.  In response, I now create a single global
variable called Optizelle, which can then be initialized once.  Inside
of the mex file, I grab this global variable, which gives access to the
structures I need, namely, the enumerated types.  Grabbing a global
variable still has some associated cost, so I actually cache this on the
mex side with some static variables.  Now, when all of that is said and
done, the new wrappers are still slower than the old.  There are a
couple of reasons why.  First, the old version of the Matlab wrappers
(not in this branch), take all the information they need from Matlab and
then do a one-shot optimization.  Meaning, features such as the state
manipulator do not exist.  When we add the state manipulator, or, for
that matter, the ability to return the entire state, we run into some
problems.  Basically, Matlab lays its memory out differently than C++,
so we can't return the state structure directly.  Either we need to
translate the state into something that Matlab understands, like a
Matlab struct, or we need to return a blob with the raw inforamtion and
a bunch of functions that can manipulate this blob.  For ease of access
to the user, and my own sanity, I went with the former option.  Hence,
we have a situation where we have the C++ state and the Matlab state and
translation routines that go back and forth between them.  These
routines absolutely have a computational cost.  In Python, because it's
memory management does a better job, we can sometimes short cut this
translation and provide shared references to different objects.  In
Matlab, we can not do this.  Hence, we end up copying more memory in
Matlab, which again has a cost.  Now, the question becomes: how
expensive is this?  For really, really small problems like Rosenbrock,
I'm getting a slow down from 1.08ish seconds to 1.82ish seconds.
However, for really large problems that I have lying around, I'm getting
a slow down from 500ish-515ish seconds, which is about 3%.  I ran a
number of tests to see whether or not the StateManipulator, which has to
translate both to and from Matlab was the largest culprit.  Removing the
statemanipulator translations on Rosenbrock, we go from 1.8s down to
1.4ish.  Hence, the StateManipulator has a cost.  However, when I ran on
the larger 500 secondish problem, removing the state manipulator did
basically nothing.  There was a second or two difference, but everything
was really a wash.  As such, on larger problems, the StateManipulator
doesn't seem to be a problem.  If it were, we could speed things up
slightly by having the user return a flag as to whether or not the
StateManipulator did anything.  If it didn't, we could skip the
translation back into C++.  With that modification, the Rosenbrock
example took 1.65ish seconds.  However, again, with the larger 500
secondish problem, there wasn't really any difference.  So, for the
moment, I'm not going to go down that path.  There are some other
differences that could be making a difference that I have not
investigated.  For example, in the old code, we stored the pointers to
the algebra like axpy, innr, copy, etc. directly.  In the new code, we
store the pointer to the algebraic object VS, which contains all of
these functions.  Then, when we want to call a function, we have to grab
the appropriate field.  This may be causing a slowdown that could be
eliminated if we go back to the old approach.  This also affects things
like the Messaging object ScalarValuedFunctions and
VectorValuedFunctions.  Mostly, I didn't really want to have to do this
because it makes the code much more verbose, but it is something to keep
in mind later.  In any case, with these changes, I have also updated the
examples so that they run properly.